Correction to avoid the problem of having the console flag set before the options are read, or the code determines if a console is being used.
Current implementation is split between using the "-noconsole"
flag to specify that the console is not being used, vs. detecting
that there is no slave interpreter.  The latter method is
probably more robust, assuming that there is no other way to
set up magic as a slave interpreter (but there probably is).